+++ tests/unit_tests/testcases/app/test_cryptoservice.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <QObject>
+
+class TestCryptoService : public QObject {
+    Q_OBJECT
+
+   private slots:
+    void testMigrateEncryptedEmptySecrets();
+    void testPreserveUndecryptableSecrets();
+};